comit

+++ src/main/java/pro/sky/telegrambot/repository/NofiticationRepository.java
@@ -5,6 +5,6 @@ import pro.sky.telegrambot.model.NofiticationTask;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface NofiticationRepository extends JpaRepository<NofiticationTask,Long> {
+public interface NofiticationRepository extends JpaRepository<NofiticationTask,Integer> {
 
 }
